requirements.txt
astroid==2.3.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
bootstrap4==0.1.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.4
django-crispy-forms==1.8.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
entrypoints==0.3
httpie==1.0.3
idna==2.8
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.6.0
nbformat==4.4.0
nose==1.3.7
notebook==6.0.1
numpy==1.17.1
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
Pygments==2.4.2
pylint==2.4.2
pylint-django==2.0.11
pylint-plugin-utils==0.6
pyparsing==2.4.2
pyrsistent==0.15.4
pyserial==3.4
python-dateutil==2.8.0
pytz==2019.2
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.4
requests==2.22.0
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
sympy==1.4
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.6
virtualenv==16.7.4
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
